+++ lib/prism_hub/domain/capabilities.rb
@@ -7,12 +7,14 @@ module PrismHub
       PUBLICATIONS_VALIDATE = "publications:validate".freeze
       PUBLICATIONS_PUBLISH = "publications:publish".freeze
       ACTORS_RESOLVE = "actors:resolve".freeze
+      ACTORS_ONBOARD = "actors:onboard".freeze
 
       ALL = [
         CHANNELS_READ,
         PUBLICATIONS_VALIDATE,
         PUBLICATIONS_PUBLISH,
-        ACTORS_RESOLVE
+        ACTORS_RESOLVE,
+        ACTORS_ONBOARD
       ].freeze
     end
   end